Redux set up for applicant profile.
